requirements.txt
Flask==1.1.1
Flask-SocketIO==4.2.1
eventlet==0.25.1
Flask-Bootstrap==3.3.7.1
ogn-client==0.9.4